등록 구현
