+++ 4기 백민기/src/components/NewBtn.js
@@ -1,6 +1,10 @@
-export default function NewBtn({ $target }) {
+export default function NewBtn({ $target, onCreate }) {
     const $newBtn = document.createElement('button')
     $newBtn.textContent = '+ New Button'
     $newBtn.className = 'newBtn'
     $target.appendChild($newBtn)
+
+    $newBtn.addEventListener('click', () => {
+        onCreate()
+    })
 }
